cv_project/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-multi-form-view==2.0.1
six==1.16.0
sqlparse==0.4.4
